--- Samples/WindowPlacement/ConsoleApplication1/FullScreenSample/resource.h
@@ -1,2 +1,0 @@
-
-#define IDC_FULLSCREEN 10102
