--- Samples/WindowPlacement/ConsoleApplication1/FullScreenSample/resources.rc
@@ -1,4 +1,0 @@
-
-#include "resource.h"
-
-IDC_FULLSCREEN      ICON      "full-screen.ico"
