--- Samples/WindowPlacement/ConsoleApplication1/SaveRestoreSample/resource.h
@@ -1,2 +1,0 @@
-
-#define IDC_SAVE 10101
