--- Samples/WindowPlacement/ConsoleApplication1/SaveRestoreSample/resources.rc
@@ -1,4 +1,0 @@
-
-#include "resource.h"
-
-IDC_SAVE      ICON      "save.ico"
