+++ Samples/WindowPlacement/cpp/FullScreenSample/FullScreenSample.filters
