+++ Samples/WindowPlacement/cpp/FullScreenSample/PropertySheet.props
