+++ Samples/WindowPlacement/cpp/SaveRestoreSample/PropertySheet.props
