+++ Samples/WindowPlacement/cpp/WindowPlacement.sln
