--- Samples/WindowPlacement/cpp/WindowPlacementSamples.sln
@@ -1,40 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "FullScreenSample", "FullScreenSample\FullScreenSample.vcxproj", "{6E745655-513E-4713-B3AB-D6D3F62D7734}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SaveRestoreSample", "SaveRestoreSample\SaveRestoreSample.vcxproj", "{92E6DB91-D852-412E-BD45-2EB168A1D090}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Debug|x64.ActiveCfg = Debug|x64
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Debug|x64.Build.0 = Debug|x64
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Debug|x86.ActiveCfg = Debug|Win32
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Debug|x86.Build.0 = Debug|Win32
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Release|x64.ActiveCfg = Release|x64
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Release|x64.Build.0 = Release|x64
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Release|x86.ActiveCfg = Release|Win32
-		{6E745655-513E-4713-B3AB-D6D3F62D7734}.Release|x86.Build.0 = Release|Win32
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Debug|x64.ActiveCfg = Debug|x64
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Debug|x64.Build.0 = Debug|x64
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Debug|x86.ActiveCfg = Debug|Win32
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Debug|x86.Build.0 = Debug|Win32
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Release|x64.ActiveCfg = Release|x64
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Release|x64.Build.0 = Release|x64
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Release|x86.ActiveCfg = Release|Win32
-		{92E6DB91-D852-412E-BD45-2EB168A1D090}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F71C6299-3377-464E-A9B6-60CE7E909160}
-	EndGlobalSection
-EndGlobal
