+++ Samples/WindowPlacement/cpp/inc/CurrentMonitorTopology.h
