+++ Samples/WindowPlacement/cpp/inc/MiscUser32.h
