+++ Samples/WindowPlacement/cpp/inc/MonitorData.h
