+++ Samples/WindowPlacement/cpp/inc/PlacementEx.h
