+++ Samples/WindowPlacement/cpp/inc/RegistryHelpers.h
