+++ Samples/WindowPlacement/cpp/inc/User32Utils.h
