+++ Samples/WindowPlacement/cpp/inc/VirtualDesktopIds.h
