+++ Samples/WindowPlacement/cpp/inc/WindowActions.h
